Rmove the manager
