+++ src/Files.App/Data/Models/ItemViewModel.cs
@@ -140,7 +140,7 @@ namespace Files.App.Data.Models
 
 			var isLibrary = false;
 			string? name = null;
-			if (App.LibraryManager.TryGetLibrary(value, out LibraryLocationItem library))
+			if (App.WindowsLibraryService.TryGetLibrary(value, out LibraryLocationItem library))
 			{
 				isLibrary = true;
 				name = library.Text;
@@ -1384,7 +1384,7 @@ namespace Files.App.Data.Models
 
 				if (path.ToLowerInvariant().EndsWith(ShellLibraryItem.EXTENSION, StringComparison.Ordinal))
 				{
-					if (App.LibraryManager.TryGetLibrary(path, out LibraryLocationItem library) && !library.IsEmpty)
+					if (App.WindowsLibraryService.TryGetLibrary(path, out LibraryLocationItem library) && !library.IsEmpty)
 					{
 						var libItem = new LibraryItem(library);
 						foreach (var folder in library.Folders)
